# openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "video-intelligence-platform",
                 "description":  "Local Video Intelligence Platform - GPU-constrained video analysis",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/videos":  {
                                  "post":  {
                                               "tags":  [
                                                            "videos"
                                                        ],
                                               "summary":  "Upload Video",
                                               "description":  "Upload a video for processing.\n\nAccepts video files up to 500MB and 10 minutes duration.\nProcessing happens asynchronously.",
                                               "operationId":  "upload_video_videos_post",
                                               "requestBody":  {
                                                                   "required":  true,
                                                                   "content":  {
                                                                                   "multipart/form-data":  {
                                                                                                               "schema":  {
                                                                                                                              "$ref":  "#/components/schemas/Body_upload_video_videos_post"
                                                                                                                          }
                                                                                                           }
                                                                               }
                                                               },
                                               "responses":  {
                                                                 "202":  {
                                                                             "description":  "Successful Response",
                                                                             "content":  {
                                                                                             "application/json":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/VideoUploadResponse"
                                                                                                                                 }
                                                                                                                  }
                                                                                         }
                                                                         },
                                                                 "422":  {
                                                                             "description":  "Validation Error",
                                                                             "content":  {
                                                                                             "application/json":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                 }
                                                                                                                  }
                                                                                         }
                                                                         }
                                                             }
                                           },
                                  "get":  {
                                              "tags":  [
                                                           "videos"
                                                       ],
                                              "summary":  "List Videos",
                                              "description":  "List all videos with pagination.",
                                              "operationId":  "list_videos_videos_get",
                                              "parameters":  [
                                                                 {
                                                                     "name":  "skip",
                                                                     "in":  "query",
                                                                     "required":  false,
                                                                     "schema":  {
                                                                                    "type":  "integer",
                                                                                    "minimum":  0,
                                                                                    "default":  0,
                                                                                    "title":  "Skip"
                                                                                }
                                                                 },
                                                                 {
                                                                     "name":  "limit",
                                                                     "in":  "query",
                                                                     "required":  false,
                                                                     "schema":  {
                                                                                    "type":  "integer",
                                                                                    "maximum":  100,
                                                                                    "minimum":  1,
                                                                                    "default":  20,
                                                                                    "title":  "Limit"
                                                                                }
                                                                 }
                                                             ],
                                              "responses":  {
                                                                "200":  {
                                                                            "description":  "Successful Response",
                                                                            "content":  {
                                                                                            "application/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/VideoListResponse"
                                                                                                                                }
                                                                                                                 }
                                                                                        }
                                                                        },
                                                                "422":  {
                                                                            "description":  "Validation Error",
                                                                            "content":  {
                                                                                            "application/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                }
                                                                                                                 }
                                                                                        }
                                                                        }
                                                            }
                                          }
                              },
                  "/videos/{video_id}":  {
                                             "get":  {
                                                         "tags":  [
                                                                      "videos"
                                                                  ],
                                                         "summary":  "Get Video",
                                                         "description":  "Get video details by ID.",
                                                         "operationId":  "get_video_videos__video_id__get",
                                                         "parameters":  [
                                                                            {
                                                                                "name":  "video_id",
                                                                                "in":  "path",
                                                                                "required":  true,
                                                                                "schema":  {
                                                                                               "type":  "string",
                                                                                               "format":  "uuid",
                                                                                               "title":  "Video Id"
                                                                                           }
                                                                            }
                                                                        ],
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "Successful Response",
                                                                                       "content":  {
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/VideoResponse"
                                                                                                                                           }
                                                                                                                            }
                                                                                                   }
                                                                                   },
                                                                           "422":  {
                                                                                       "description":  "Validation Error",
                                                                                       "content":  {
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                           }
                                                                                                                            }
                                                                                                   }
                                                                                   }
                                                                       }
                                                     },
                                             "delete":  {
                                                            "tags":  [
                                                                         "videos"
                                                                     ],
                                                            "summary":  "Delete Video",
                                                            "description":  "Delete a video and all associated data.",
                                                            "operationId":  "delete_video_videos__video_id__delete",
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "video_id",
                                                                                   "in":  "path",
                                                                                   "required":  true,
                                                                                   "schema":  {
                                                                                                  "type":  "string",
                                                                                                  "format":  "uuid",
                                                                                                  "title":  "Video Id"
                                                                                              }
                                                                               }
                                                                           ],
                                                            "responses":  {
                                                                              "204":  {
                                                                                          "description":  "Successful Response"
                                                                                      },
                                                                              "422":  {
                                                                                          "description":  "Validation Error",
                                                                                          "content":  {
                                                                                                          "application/json":  {
                                                                                                                                   "schema":  {
                                                                                                                                                  "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                              }
                                                                                                                               }
                                                                                                      }
                                                                                      }
                                                                          }
                                                        }
                                         },
                  "/videos/{video_id}/jobs":  {
                                                  "get":  {
                                                              "tags":  [
                                                                           "videos"
                                                                       ],
                                                              "summary":  "Get Video Jobs",
                                                              "description":  "Get all processing jobs for a video.",
                                                              "operationId":  "get_video_jobs_videos__video_id__jobs_get",
                                                              "parameters":  [
                                                                                 {
                                                                                     "name":  "video_id",
                                                                                     "in":  "path",
                                                                                     "required":  true,
                                                                                     "schema":  {
                                                                                                    "type":  "string",
                                                                                                    "format":  "uuid",
                                                                                                    "title":  "Video Id"
                                                                                                }
                                                                                 }
                                                                             ],
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "Successful Response",
                                                                                            "content":  {
                                                                                                            "application/json":  {
                                                                                                                                     "schema":  {
                                                                                                                                                    "type":  "array",
                                                                                                                                                    "items":  {
                                                                                                                                                                  "$ref":  "#/components/schemas/JobResponse"
                                                                                                                                                              },
                                                                                                                                                    "title":  "Response Get Video Jobs Videos  Video Id  Jobs Get"
                                                                                                                                                }
                                                                                                                                 }
                                                                                                        }
                                                                                        },
                                                                                "422":  {
                                                                                            "description":  "Validation Error",
                                                                                            "content":  {
                                                                                                            "application/json":  {
                                                                                                                                     "schema":  {
                                                                                                                                                    "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                }
                                                                                                                                 }
                                                                                                        }
                                                                                        }
                                                                            }
                                                          }
                                              },
...
              },
    "components":  {
                       "schemas":  {
                                       "Body_upload_video_videos_post":  {
                                                                             "properties":  {
                                                                                                "file":  {
                                                                                                             "type":  "string",
                                                                                                             "format":  "binary",
                                                                                                             "title":  "File",
                                                                                                             "description":  "Video file to upload"
                                                                                                         }
                                                                                            },
                                                                             "type":  "object",
                                                                             "required":  [
                                                                                              "file"
                                                                                          ],
                                                                             "title":  "Body_upload_video_videos_post"
                                                                         },
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
...
                                   }
                   }
}